restruct code
