+++ controller/kvtop.py
@@ -4,7 +4,7 @@ import os
 import time
 from typing import List, Optional, Union
 
-from kvcached.utils import SHM_DIR, MemInfoStruct, RwLockedShm, get_ipc_name
+from utils import SHM_DIR, MemInfoStruct, RwLockedShm, get_ipc_name
 
 
 def _format_size(num_bytes: int) -> str:
